feat(refactor): use webpack instead of gulp, installed vue 3 for website
