--- scss/magic-mouse.scss
@@ -1,52 +1,0 @@
-body {
-  #magicMouseCursor {
-    position: fixed;
-    width: 35px;
-    height: 35px;
-    border: 1px solid #fff;
-    border-radius: 50%;
-    z-index: 9999;
-    left: 0;
-    top: 0;
-    transition: transform 0.07s, width 0.3s, height 0.3s;
-    pointer-events: none;
-    &.cursor-square {
-      border-radius: 0;
-    }
-  }
-  #magicPointer {
-    height: 5px;
-    width: 5px;
-    top: 0;
-    left: 0;
-    position: fixed;
-    background: #fff;
-    border-radius: 50%;
-    pointer-events: none;
-    transition: background 0.2s, width 0.2s, height 0.2s, box-shadow 0.2s;
-    &.is-hover {
-      background: red;
-    }
-    &.pointer-blur {
-      height: 50px;
-      width: 50px;
-      background: none;
-      border: 1px solid #fff;
-      box-shadow: 0px 0px 15px -5px rgba(255, 255, 255, 1);
-    }
-    &.pointer-overlay {
-      height: 50px;
-      width: 50px;
-      mix-blend-mode: difference;
-      box-shadow: 0px 0px 15px -5px rgba(255, 255, 255, 1);
-    }
-  }
-  .magic-hover {
-    transition: all 0.2s;
-    &:hover {
-      cursor: none;
-
-    }
-
-  }
-}
