--- scss/website.scss
@@ -1,135 +1,0 @@
-$primary_color: #22252c;
-
-
-/* desert scheme ported from vim to google prettify */
-pre.prettyprint {
-  display: block; background-color: #4A4D55;
-  border:solid 12px #1F2128!important;
-  padding:20px 0 20px 30px!important;
-}
-pre .nocode { background-color: none; color: #000 }
-pre .str { color: #ffa0a0 } /* string  - pink */
-pre .kwd { color: #f0e68c; font-weight: bold }
-pre .com { color: #87ceeb } /* comment - skyblue */
-pre .typ { color: #98fb98 } /* type    - lightgreen */
-pre .lit { color: #cd5c5c } /* literal - darkred */
-pre .pun { color: #fff }    /* punctuation */
-pre .pln { color: #fff }    /* plaintext */
-pre .tag { color: #f0e68c; font-weight: bold } /* html/xml tag    - lightyellow */
-pre .atn { color: #bdb76b; font-weight: bold } /* attribute name  - khaki */
-pre .atv { color: #ffa0a0 } /* attribute value - pink */
-pre .dec { color: #98fb98 } /* decimal         - lightgreen */
-
-/* Specify class=linenums on a pre to get line numbering */
-ol.linenums { margin-top: 0; margin-bottom: 0; color: #AEAEAE } /* IE indents via margin-left */
-li.L0,li.L1,li.L2,li.L3,li.L5,li.L6,li.L7,li.L8 { list-style-type: none }
-/* Alternate shading for lines */
-li.L1,li.L3,li.L5,li.L7,li.L9 { }
-
-@media print {
-pre.prettyprint { background-color: none }
-pre .str, code .str { color: #060 }
-pre .kwd, code .kwd { color: #006; font-weight: bold }
-pre .com, code .com { color: #600; font-style: italic }
-pre .typ, code .typ { color: #404; font-weight: bold }
-pre .lit, code .lit { color: #044 }
-pre .pun, code .pun { color: #440 }
-pre .pln, code .pln { color: #000 }
-pre .tag, code .tag { color: #006; font-weight: bold }
-pre .atn, code .atn { color: #404 }
-pre .atv, code .atv { color: #060 }
-}
-
-
-.btn{
-  border-radius: 22px;
-  padding:10px 25px;
-  font-size: 13px;
-  font-weight: bold;
-  text-transform: uppercase;
-  position: relative;
-  &:focus{
-    box-shadow: none;
-  }
-  &.btn-mm-dark{
-    background: #4F4F4F;
-    color: #fff;
-  }
-  &.btn-mm-light{
-    background: #EB5757;
-    color: #fff;
-  }
-  &.github-icon{
-    padding-left: 44px;
-    &::before{
-      content:'';
-      position: absolute;
-      width: 31px;
-      height: 30px;
-      background: url('../images/github-icon.svg');
-      left: 6px;
-      top: 50%;
-      transform: translateY(-50%);
-    }
-  }
-}
-
-body {
-  background: $primary_color;
-  color: #fff;
-  font-family: 'Quicksand', sans-serif;
-  .container-fluid{
-    nav{
-      background:none;
-      .navbar-collapse{
-        justify-content: flex-end;
-        .navbar-nav{
-          li{
-            padding:10px;
-            a{
-              font-size: 18px;
-              font-weight: 400;
-            }
-          }
-        }
-      }
-    }
-  }
-  .main-body{
-    margin-top: 200px;
-    .code-section{
-      .try-section{
-        text-align: center;
-        select{
-          display: inline-block;
-          max-width:200px;
-          height: 40px;
-          margin-right: 20px;
-          outline: none;
-          &:focus{
-            outline: 0;
-            border:none;
-            box-shadow: none;
-          }
-        }
-      }
-    }
-    .text-section{
-      h1{
-        font-family: 'Alfa Slab One', cursive;
-        font-style: normal;
-        font-weight: normal;
-        font-size: 26px;
-        line-height: 36px;
-      }
-      p{
-        font-weight: 500;
-        font-size: 18px;
-        line-height: 22px;
-      }
-      .btn-section{
-        margin-top: 115px;
-      }
-    }
-  }
-}
